c++: Handle RAW_DATA_CST in add_list_candidates [PR118532]
This is the second bug discovered today with the
https://gcc.gnu.org/pipermail/gcc-patches/2025-January/673945.html
hack but then turned into proper testcases where embed-2[23].C FAILed
since introduction of optimized #embed support and the others when
optimizing large C++ initializers using RAW_DATA_CST.

The add_list_candidates problem is the same as with
make_tree_vector_from_ctor, unfortunately it can't call that
function because it can have those additional artificial arguments
that need to be pushed earlier.
When working on the patch, I've also noticed an error where we didn't
know how to dump RAW_DATA_CST, so I've added support for that too.

2025-01-21  Jakub Jelinek  <jakub@redhat.com>

	PR c++/118532
	* call.cc (add_list_candidates): Handle RAW_DATA_CST among init_list
	elts.
	* error.cc (dump_expr_init_vec): Handle RAW_DATA_CST among v elts.

	* g++.dg/cpp/embed-22.C: New test.
	* g++.dg/cpp/embed-23.C: New test.
	* g++.dg/cpp0x/pr118532.C: New test.
	* g++.dg/cpp2a/explicit20.C: New test.