testsuite: Fix test failing with -fimplicit-constexpr [PR118277]
While testing an unrelated C++ patch with "make check-c++-all", I
noticed that r15-6760-g38a13ea4117b96 added a test case that fails with
-fimplicit-constexpr.

The problem is that this test unconditionally expects an error stating
that a non-constexpr function is called, but that function is
auto-magically constexpr'd under -fimplicit-constexpr.

As suggested by Jakub, this patch simply passes -fno-implicit-constexpr
in that test.

	PR c++/118277

gcc/testsuite/ChangeLog:

	* g++.dg/cpp1z/constexpr-asm-5.C: Pass -fno-implicit-constexpr.