c++: Don't ICE in build_class_member_access_expr during error recovery [PR118225]
The invalid case in this PR trips on an assertion in
build_class_member_access_expr that build_base_path would never return
an error_mark_node, which is actually incorrect if the object involves a
tree with an error_mark_node DECL_INITIAL, like here.

This patch changes the assert to not fire if an error has been reported.

	PR c++/118225

gcc/cp/ChangeLog:

	* typeck.cc (build_class_member_access_expr): Let errors that
	that have been reported go through.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp0x/constexpr-ice21.C: New test.